Various bug fixes and code cleanup.
- Fixes issue where various latency metrics were scaled incorrectly when the exporter was running as an agent/in-process.
- Fixes quoted argument parsing code for agent.
- Improves the caching implementation of MetricFamiles, expecailly those that contain multiple values (such as Summaries and Histograms)
