+++ agent/src/main/java/com/zegelin/agent/AgentArgumentParser.java
@@ -49,6 +49,7 @@ public final class AgentArgumentParser {
 
                     if (c == '"') {
                         stateStack.push(State.QUOTED);
+                        continue;
                     }
 
                     break;
@@ -56,10 +57,12 @@ public final class AgentArgumentParser {
                 case QUOTED:
                     if (c == '"') {
                         stateStack.pop();
+                        continue;
                     }
 
                     if (c == '\\') {
                         stateStack.push(State.ESCAPED);
+                        continue;
                     }
 
                     break;
