+++ common/src/main/java/com/zegelin/cassandra/exporter/netty/HttpHandler.java
@@ -10,8 +10,8 @@ import com.zegelin.netty.Resources;
 import com.zegelin.cassandra.exporter.Harvester;
 import com.zegelin.prometheus.domain.Labels;
 import com.zegelin.prometheus.domain.MetricFamily;
-import com.zegelin.prometheus.exposition.JsonFormatChunkedInput;
-import com.zegelin.prometheus.exposition.TextFormatChunkedInput;
+import com.zegelin.prometheus.exposition.json.JsonFormatChunkedInput;
+import com.zegelin.prometheus.exposition.text.TextFormatChunkedInput;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.ByteBufUtil;
 import io.netty.channel.ChannelFuture;
