+++ common/src/main/java/com/zegelin/prometheus/domain/SummaryMetricFamily.java
@@ -1,5 +1,7 @@
 package com.zegelin.prometheus.domain;
 
+import com.google.common.collect.ImmutableList;
+
 import java.util.List;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
@@ -7,10 +9,10 @@ import java.util.stream.Stream;
 
 public class SummaryMetricFamily extends MetricFamily<SummaryMetricFamily.Summary> {
     public SummaryMetricFamily(final String name, final String help, final Stream<Summary> metrics) {
-        super(name, help, metrics);
+        this(name, help, () -> metrics);
     }
 
-    private SummaryMetricFamily(final String name, final String help, final Supplier<Stream<Summary>> metricsStreamSupplier) {
+    SummaryMetricFamily(final String name, final String help, final Supplier<Stream<Summary>> metricsStreamSupplier) {
         super(name, help, metricsStreamSupplier);
     }
 
@@ -20,7 +22,7 @@ public class SummaryMetricFamily extends MetricFamily<SummaryMetricFamily.Summar
     }
 
     @Override
-    public MetricFamily<Summary> cache() {
+    public SummaryMetricFamily cachedCopy() {
         final List<Summary> metrics = metrics().collect(Collectors.toList());
 
         return new SummaryMetricFamily(name, help, metrics::stream);
@@ -29,14 +31,14 @@ public class SummaryMetricFamily extends MetricFamily<SummaryMetricFamily.Summar
     public static class Summary extends Metric {
         public final float sum;
         public final float count;
-        public final Stream<Interval> quantiles;
+        public final Iterable<Interval> quantiles;
 
-        public Summary(final Labels labels, final float sum, final float count, final Stream<Interval> quantiles) {
+        public Summary(final Labels labels, final float sum, final float count, final Iterable<Interval> quantiles) {
             super(labels);
 
             this.sum = sum;
             this.count = count;
-            this.quantiles = quantiles;
+            this.quantiles = ImmutableList.copyOf(quantiles);
         }
     }
 }
