+++ common/src/main/resources/com/zegelin/prometheus/exposition/text/banner.txt
